[Refactoring] int type to NSInteger

+++ LYCategory/Classes/_Foundation/_Date/NSDate+Formatter.h
@@ -49,7 +49,7 @@
 
 #pragma mark WEEKDAY
 
-- (int)weekdayNumber;
+- (NSInteger)weekdayNumber;
 
 - (int)weekdayNumberPRC;
